stream_processing/requirements.txt
confluent-kafka
avro 
urllib3 
requests
fastavro
google-cloud-storage
aiohttp
aiofiles